+++ examples/github_search/common_github_search/lib/src/github_search_bloc/github_search_bloc.dart
@@ -16,7 +16,7 @@ class GithubSearchBloc extends Bloc<GithubSearchEvent, GithubSearchState> {
     Stream<GithubSearchEvent> events,
     Stream<GithubSearchState> Function(GithubSearchEvent event) next,
   ) {
-    return (events as Observable<GithubSearchEvent>)
+    return events
         .debounceTime(
           Duration(milliseconds: 300),
         )