+++ packages/bloc_test/lib/src/emits_exactly.dart
@@ -25,13 +25,26 @@ import 'package:test/test.dart';
 ///   await emitsExactly(bloc, [isA<StateA>(), isA<StateB>()]);
 /// });
 /// ```
+///
+/// [emitsExactly] also takes an optional [duration] which is useful in cases
+/// where the [bloc] is using `debounceTime` or other similar operators.
+///
+/// ```dart
+/// test('emits [0, 1] when CounterEvent.increment is added', () async {
+///   final bloc = CounterBloc();
+///   bloc.add(CounterEvent.increment);
+///   await emitsExactly(bloc, [0, 1], duration: const Duration(milliseconds: 300));
+/// });
+/// ```
 Future<void> emitsExactly<B extends Bloc<dynamic, State>, State>(
   B bloc,
-  Iterable expected,
-) async {
+  Iterable expected, {
+  Duration duration,
+}) async {
   assert(bloc != null);
   final states = <State>[];
   final subscription = bloc.listen(states.add);
+  if (duration != null) await Future.delayed(duration);
   await bloc.close();
   expect(states, expected);
   await subscription.cancel();