+++ packages/bloc_test/test/emits_exactly_test.dart
@@ -125,6 +125,68 @@ void main() {
       });
     });
 
+    group('DebounceCounterBloc', () {
+      test('emits [0, 1] when CounterEvent.increment is added', () async {
+        final bloc = DebounceCounterBloc();
+        bloc.add(DebounceCounterEvent.increment);
+        await emitsExactly(bloc, [0, 1],
+            duration: const Duration(milliseconds: 305));
+      });
+
+      test('fails if bloc does not emit all states', () async {
+        try {
+          await emitsExactly(DebounceCounterBloc(), [0, 1]);
+          fail('should throw');
+        } on TestFailure catch (error) {
+          expect(
+              error.message,
+              'Expected: [0, 1]\n'
+              '  Actual: [0]\n'
+              '   Which: shorter than expected at location [1]\n'
+              '');
+        }
+      });
+
+      test('fails if bloc does not emit correct states', () async {
+        try {
+          final bloc = DebounceCounterBloc();
+          bloc.add(DebounceCounterEvent.increment);
+          await emitsExactly(bloc, [0, 2],
+              duration: const Duration(milliseconds: 305));
+          fail('should throw');
+        } on TestFailure catch (error) {
+          expect(
+            error.message,
+            'Expected: [0, 2]\n'
+            '  Actual: [0, 1]\n'
+            '   Which: was <1> instead of <2> at location [1]\n'
+            '',
+          );
+        }
+      });
+
+      test('fails if expecting extra states', () async {
+        try {
+          final bloc = DebounceCounterBloc();
+          bloc.add(DebounceCounterEvent.increment);
+          await emitsExactly(
+            bloc,
+            [0, 1, 2],
+            duration: const Duration(milliseconds: 305),
+          );
+          fail('should throw');
+        } on TestFailure catch (error) {
+          expect(
+            error.message,
+            'Expected: [0, 1, 2]\n'
+            '  Actual: [0, 1]\n'
+            '   Which: shorter than expected at location [2]\n'
+            '',
+          );
+        }
+      });
+    });
+
     group('MultiCounterBloc', () {
       test('emits [0, 1, 2] when CounterEvent.increment is added', () async {
         final bloc = MultiCounterBloc();